State: all tests are passing after re-structuring
