+++ docs/source/proj_structure.rst
@@ -1,0 +1,11 @@
+
+- gbox
+    - apps/
+    - misc.py
+    - shapes.py
+    - elements.py
+    - gbox.py
+    - utils.py
+
+``apps`` should sit at the top of all the modules and ``utilities`` should be at the base.
+
